2prueba_solution: AWS deployment success
Add Terraform files to automate the creation of free ec2 instances to deploy the frontend backend and db containers of the application.
Instance ec2 with security group and keys to access via ssh.
Initial instance setup: install docker-compose git docker repository and save the public ip of the instance to use it as environment variable